KL-119/refactor: separate UserService
